separate modules step 1
