+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cervus",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "author": "Michał Budzyński <michal@virtualdesign.pl>",
   "repository": "git@github.com:michalbe/cervus.git",
   "scripts": {
