+++ src/cervus.js
@@ -1,0 +1,23 @@
+import { Game } from './core/game.js';
+import { Entity } from './core/entity.js';
+import { math } from './core/math.js';
+export { Entity };
+export { Game };
+export { math };
+
+import { Basic } from './materials/basic.js';
+import { Phong } from './materials/phong.js';
+export { Basic };
+export { Phong };
+
+import { defaults } from './misc/defaults.js';
+import { utils } from './misc/utils.js';
+export { defaults };
+export { utils };
+
+import { Box } from './shapes/box.js';
+import { Plane } from './shapes/plane.js';
+import { Sphere } from './shapes/sphere.js';
+export { Box };
+export { Plane };
+export { Sphere };
