+++ src/main.js
@@ -1,12 +1,13 @@
 import { Game } from './core/game.js';
 import { Entity } from './core/entity.js';
 import { math } from './core/math';
-
+import { materials } from './materials/materials.js';
 import { shapes } from './shapes/shapes.js';
 
 export {
   Game,
   Entity,
   shapes,
-  math
+  math,
+  materials
 }
