+++ src/materials/materials.js
@@ -1,5 +1,5 @@
-import { Basic } from './basic/basic.js'
-import { Phong } from './phong/phong.js'
+import { Basic } from './basic.js'
+import { Phong } from './phong.js'
 // import { ShadowPhong } from './shadow-phong/shadow-phong.js'
 
 const materials = {
